break network to a separate file
