+++ test/test-driver.js
@@ -1,0 +1,36 @@
+/**
+ * Copyright 2018 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+   * @param {!EventEmitter} emitter
+   * @param {string} eventName
+   * @param {number=} eventCount
+   * @return {!Promise<!Object>}
+   */
+module.exports.waitForEvents = function waitForEvents(emitter, eventName, eventCount = 1) {
+  let fulfill;
+  const promise = new Promise(x => fulfill = x);
+  emitter.on(eventName, onEvent);
+  return promise;
+
+  function onEvent(event) {
+    --eventCount;
+    if (eventCount)
+      return;
+    emitter.removeListener(eventName, onEvent);
+    fulfill(event);
+  }
+};
